profiles/base: Add patch for networkd IPv6 fix
The upstream issue is systemd/systemd#3879 and this bug has annoyed me
for a while now. I've also opened a pull request (NixOS/systemd#12) to
the NixOS fork of systemd, but this is pending since almost a month now
and the bug is still annoying as fuck, so let's patch it in vuizvui.

I'm not yet porting this to other machines yet, because I want to test
whether this actually really solves my problem (even though it's exactly
what the upstream bug describes) or whether we need to patch something
else as well.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>
Cc: @devhell (uses networkd as well)