modules/aszlig: Make vim module a plain package
I don't use anything that's machine specific within my Vim
configuration (and even if, we can pass it via the callPackage
arguments) so it's kinda pointless that it's a module instead of a plain
package (override).

This makes it also easier to nix-build the package without the need to
go through the module system.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>