pkgs/vim: Switch Haxe plugin to vaxe
The Haxe plugin(s) I've used so far (vim-haxe and vim-haxe-syntax) are
no longer actively maintained since 2014.

On the other hand "vaxe" is based on the work on vim-haxe and
vim-haxe-syntax, is actively maintained and supports newer language
features of Haxe as well.

I've patched out syntax highlighting for ',' and ';', because I really
get eye cancer with this and reminds me a bit about the annoying bold
colons when using nickname completion back in the days where BitchX
users were widely seen on IRC.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>