pkgs/vim: Be more agressive with including plugins
Originally I only wanted to make sure the ftplugin gets loaded before
the syntax file, but while at it I thought just prepending/appending
stuff to the runtimepath is not enough for me.

So now my version of Vim has all the plugins directly in the standard
directories just as if they came with upstream.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>