pkgs/vim: Add plugin for editing CSV files
I happen to work with CSV files quite a bit lately but it's a major
nuisance doing so with a normal text editor and I don't want to use a
full-blown spreadsheet programm just for a few CSV files or even write a
script every time I need to edit only a tiny bit.

So this plugin solves that problem for me.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>