pkgs/vim: Clean up autocommands
I was using set for a bunch of these, but these autocommands are only
used for single buffers so let's actually make sure they are set only
there by using setlocal.

In addition to that I've corrected usage of '==' to use '==#', because
'==' actually depends on user settings whether it's case sensitive or
not (set ignorecase).

Signed-off-by: aszlig <aszlig@redmoonstudios.org>