machines: Unify most options for brawndo and tyree
Both machines are using Plasma along with a few common KDE applications,
so let's actually be more aggressive about deduplicating the options,
because after all whenever these machines start to deviate more from
each other, we can still either override those options or move them out.

I've also cleaned up a few packages, so instead of having mpv *and* vlc
we now have mpv and bomi for both machines. The latter is mostly about
figuring out whether it's actually a good GUI video player, as it is
basically a front-end for mpv.

With this unification this means that some packages which are generally
useful, such as okular and gwenview are now not only available to tyree
but brawndo as well.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>
Cc: @dwenola, @BrokkoliBerta