games/fetch-itch: Allow to pass a version
So far we had a similar "experience" with itch.io as we had with
humblebundle having upstream URLs being updated in-place and thus
whenever we're trying to get an earlier revision than the latest one,
our fetcher would fail.

However, itch.io has build IDs which even contain version string that we
can actually use in our packages as-is and they directly map to a build
ID.

This commit implements just that, so that whenever a version attribute
is passed to fetchItch, an explicit version is selected. If such a
version couldn't be found, a list of the most recent versions is
displayed, which should also make it easier for updating packages.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>