pkgs/games: Add a new buildGame function
The main functionality for this function is to gather missing
dependencies in ELF executables and shared libraries and using patchelf
to set the right RPATH.

All of the dependencies are searched based on what we have in one of the
buildInputs variables, so all we need to do is list them in there.

One thing that's still left to solve is adding libraries to the RPATH
which are only required at runtime. An example for this would be the
pulseaudio library.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>