pkgs/build-game: Introduce buildSandbox
So far creating the sandbox has been a setup hook, however it's a bit
ugly how we gathered the needed paths for the chroot file system by
recursively searching for store paths.

While I'd like to have the sandbox being built within the main
derivation, it really isn't very practical when the build takes longer
than 10 minutes.

With this implementation however the sandbox builds really fast and we
can also modify the sandbox without needing to rebuild a particular
game.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>