pkgs/sandbox: Include basic runtime paths
These paths are things such as /etc and /run but also the .Xauthority
file, which contains the cookie to connect to the X server.

What is still missing is access to the DRI libraries, which is a bit
trickier, because we need to add those store paths at runtime and we
need to also mount all of the dependencies.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>