hardware/t100ha: Update to kernel 4.14
Almost all of the issues we so far needed patches for have been
addressed upstream with the exception of backlight probing.

The reason for this is detailled in the following bug:

https://bugs.freedesktop.org/show_bug.cgi?id=96571

In summary, when the I915 driver is built into the kernel, it can't
access the PWM because it's initialized at a later stage.

What we're now doing instead is that we compile i915 as a module and
bake the necessary PWM modules directly into the kernel.

This has the advantage that we no longer need the backlight.patch,
because after all this is already a workaround and having a workaround
that doesn't require us to rebase the patch once a new kernel arrives
seems to be a better solution in my opinion.

With this update we can now revert the reintroduction of kernel 4.12
done in 45bdcd7fc7788316cbad2d7d412e421b230715dc.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>