profiles/managed: Add ark to systemPackages
This really makes sense for an end user system, so let's actually add it
to the system path. It seems that it also needs unzip and other
archivers in $PATH but I haven't actually verified if that's really the
case, because the Nix expression for ark already contains those
archivers.

Signed-off-by: aszlig <aszlig@nix.build>