pkgs: Remove the vuizvui namespace
It's a bit weird to need to write something like this:

nix-build pkgs -A vuizvui.aszlig.vim

We're already within vuizvui, and the filesystem paths are already
structured in a way so that the above is in pkgs/aszlig/vim, so the need
to add the vuizvui attribute to that seems quite pointless and
unnatural.

I have tested this by evaluating a few machines, tests and packages.

Signed-off-by: aszlig <aszlig@nix.build>