taalo-build: Rewrite as wrapper and move to pkgs
We no longer use the legacy SSH store protocol for taalo but the new
ssh-ng protocol, which makes the implementation of taalo-build a LOT
less clunky.

It also didn't make sense to have this as a NixOS module when we after
all just emit a static store path without any stuff depending on
configuration options.

The new implementation basically just wraps nix-build and nix-store -r
along with the right NIX_REMOTE variable.

With Nix 1.2 this can also be done with the new "nix build" command
using the --store option, but unfortunately "nix build" doesn't yet have
the same functionality as nix-build.

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @Profpatsch, @bendlas