tests/games/sandbox: Rename to tests/sandbox
We have excluded all tests within the games directory from being built
by Hydra, rightfully so because they're proprietary. However our sandbox
is *not* proprietary so we want to have it tested.

Besides, we might want to use that sandbox for other things rather than
just games in the future, which saves us that rename later ;-)

Signed-off-by: aszlig <aszlig@nix.build>