pkgs/sandbox: Handle mounting of regular files
While we already have support for mounting plain files, this is done on
a very specific basis, mainly the .Xauthority file.

Whenever we use bind_mount() and the file is a regular file, mounting
that file will fail. So let's actually do a stat on the file and decide
whether we want to do bind_file() or bind_mount().

I've stumbled on this because one of the store paths of the run time
dependency graph was a plain file and thus the sandbox wrapper was
unable to mount it.

Signed-off-by: aszlig <aszlig@nix.build>