pkgs/sandbox: Add handling for XDG_CACHE_HOME
We only handle XDG_DATA_HOME and XDG_CONFIG_HOME, but we've missed
XDG_CACHE_HOME. While the latter is used very rarely as it doesn't
matter a lot if it ends up within a tmpfs anyway. However if the cache
directory gets pretty large we might run out of space.

Not only do we now have proper fallbacks but this also adds tests for
all of the XDG environment variables we're using.

Signed-off-by: aszlig <aszlig@nix.build>