core/lazy-packages: Silence stdout of nix-store -r
For example if you have something like cabal2nix as a lazy package,
using "cabal2nix --shell . > default.nix" will result in the store path
of cabal2nix being present in the resulting default.nix as the first
line.

We obviously do not want that, so let's redirect stdout of nix-store -r
to /dev/null.

Signed-off-by: aszlig <aszlig@nix.build>