games/humblebundle: Add Opus Magnum
Upstream URL: http://www.zachtronics.com/opus-magnum/

For this I've ripped apart all of the bundled shared objects including
Mono itself, so we're using Mono from <nixpkgs> instead of the provided
one.

In addition to that I had to patch the intermediate language because the
game really doesn't cope very well when the data files are read-only,
which is the case in the Nix store.

Instead of replacing the CIL, I could have used an LD preloader as well,
but I think this is way less error-prone even though we had to patch a
few internal classes where we couldn't match the name properly.

Compared to this an LD preloader would involve patching several symbols
and also implement some logic to distinguish between files read from the
data directory and files read/written to XDG_DATA_HOME.

The reason why I added $HOME/Desktop to the required sandbox paths is
because the game saves gif images into that directory.

Signed-off-by: aszlig <aszlig@nix.build>