pkgs/santander: Update for Wine 3.0
The wine-no-unixfs.patch doesn't apply anymore, so we need to get a
newer version of the patch now from http://achurch.org/patch-pile/,
which is a bit different in that it adds --disable-unixfs as a flag to
configure. We don't have autoreconfHook in nativeBuildInputs, so that
flag will simply be ignored but unixfs will be disabled regardless
because the flag in config.h.in is ENABLE_UNIXFS and it's not defined by
default. The reason why I still added it to configureFlags is to prevent
enabling it by accident if autoreconfHook should be part of the wine
derivation in upstream <nixpkgs>.

In addition to that I've made disabling getuid a bit more clear and with
a comment about why we need to disable it.

Signed-off-by: aszlig <aszlig@nix.build>