services/i3/conky: Fix segfault
We're using "out_to_x = false" because we only want output to stdout,
unfortunately since conky version 1.10.7, this causes a segfault which
is tracked in upstream issue brndnmtthws/conky#454.

The patch I'm using here was submitted and merged upstream in
brndnmtthws/conky#455 and it's not yet part of a release yet.

Signed-off-by: aszlig <aszlig@nix.build>