auto-patchelf: Only search for deps with same arch
So far we only matched the file name of the dependency but not its
architecture, so if for example there is one shared object for
i686-linux and another one with the same name but for x86_64-linux,
chances are that the wrong architecture is chosen.

Now we're checking the architecture of the shared object file and only
pick it, if it matches the architecture of the file to patchelf.

Signed-off-by: aszlig <aszlig@nix.build>