pkgs/build-sandbox: Zero-pad minor Nix version
In order to do integer comparisons on the Nix version, we need to
zero-pad the minor version, so that we always have two digits.

Since the change of Nix version 1.12 to 2.0 the minor version no longer
has two digits, so we get 20 instead of 112 and when compared the former
is smaller than the latter but it has to be the opposite.

Signed-off-by: aszlig <aszlig@nix.build>