dnyarri: Override gpodder to use latest master
This is mainly to get gpodder/gpodder@c937184987431427dfcf8ac9bc098ce0ac
but there are other fixes in the current master version, so instead of
just patching the single fix, I'm heading for master directly instead.

I've also added an assertion on the version attribute so that whenever
there is a new upstream version, we can revert this.

Signed-off-by: aszlig <aszlig@nix.build>