release.nix: Exclude broken packages from set
While we now evaluate broken packages and they fail to build, this is
still not very nice because it will still break the generic channel.

So let's do this a bit differently and actually exclude these packages
without running into an eval error. We evaluate the derivation and later
check whether meta.broken is set, which has the advantage that we only
kick out packages that are *directly* marked as broken rather than
excluding packages that have a broken flag in one of their dependencies.

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @Profpatsch