core/tests: Update upstream NixOS tests
New tests as of NixOS/nixpkgs@27a4a5511dc2cf9756943458d7a752617e4ad6e3:

 acme, atd, beegfs, couchdb, dovecot, env, gitolite, grafana, graphite,
 home-assistant, initrd-network-ssh, kernel-copperhead, kernel-latest,
 kernel-lts, mesos, mutableUsers, mysqlBackup, netdata, nghttpx,
 owncloud, prometheus, radicale, statsd, sudo, switchTest, vault,
 xmonad, zookeeper

I've removed the pgjwt test, because the attribute is not reachable
anymore within nixos/release.nix as it has been commented out in
NixOS/nixpkgs@40ed22650737a444fee7883e69e852fd4b808133.

Also I didn't add the postgis test, because checking for the
configuration attribute is a bit tedious and chances are very low that
we'd have a machine in Vuizvui that uses it.

Signed-off-by: aszlig <aszlig@nix.build>