profpatsch/searx: Rebase searx-secret-key.patch
Get rid of the annoying build error again (see 8be3704c3fbe08308132d92a6
for the last time), but this time for searx 0.13.1.

The main thing that has changed is that instead of hmac.new() there is
now a helper function called new_hmac(), which the patch now uses.

I've also cleaned up some tiny bits reported by flake8 and fixed the
description of SecretAppKeyError to refer to XDG_CACHE_HOME instead of
XDG_CACHE_DIR.

In addition to the updated patch, we now need to add XDG_CACHE_HOME to a
writable directory (in this case $TMPDIR), because otherwise the
test_webapp fails with an SecretAppKeyError, as it will try to create
leading directories to XDG_CACHE_HOME.

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @Profpatsch