build-game: Properly find sourceRoot
So far we only checked for the source root by checking only one level of
directories, so for example if the source code is unpacked into
something like "a/b/c" and only "c" contains files, the previous
implementation would set the source root to be "a".

In addition we didn't do a check on the type of the contents, so for
example if the archive contains only *one* file, that *file* itself
would be set as the source root and the chdir to that will fail later.

What we do now is recursively search through directories in order to
find the last directory that does not contain *only* a subdirectory.

Signed-off-by: aszlig <aszlig@nix.build>