dnyarri: Remove gpodder derivation attrs override
I've put in an assertion back then to make sure that once gpodder 3.10.1
is released, we get a failure so we don't stay behind forever.

With NixOS/nixpkgs@4d1e72cfbb8de0d8adf2c047aad14f29eb4f77bf, not only
the version is bumped but also the icon theme is added back, so we can
drop the whole override of the attributes.

Signed-off-by: aszlig <aszlig@nix.build>