sternenseemann/logbook: Support newer Angstrong
Angstrong doesn't work anymore for OCaml versions lower that 4.03, so
let's use version 4.04.2 and fix logbook to cope with the interface
changes in Angstrong 0.7.0.

I've opened a pull request (sternenseemann/logbook#1) to fix it upstream
and also using the commit from the pull request here.

This fixes the following evaluation error:

in job 'channels.generic':
angstrom is not available for OCaml 4.02.3

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @sternenseemann