starbound: Fix preloader snprintf size
I explicitly didn't want to include a \0, but unfortunately with GCC 7 a
check is performed on the size of the format arguments and the build now
fails with:

'__builtin___snprintf_chk' output between 5 and 9 bytes into a
destination of size 4

The dereference of the str pointer gives us one byte, so the size of the
format argument will be 4 in any cases plus the 0 byte at the end. While
we don't need the 0 byte because it will be added later, it doesn't hurt
so let's use a maximum size of 5 to keep GCC 7 silent.

Signed-off-by: aszlig <aszlig@nix.build>