machines/arilou: Use correct modulesPath
Since I wrote the config for that machine, the modulesPath attribute is
actually a thing in NixOS modules, so let's use that one instead.

Besides, I really tend to like that even more restrictive restrictive
evaluation mode, because it allows us to get rid of all those references
to the Nix path.

Signed-off-by: aszlig <aszlig@nix.build>