Fix up call-network and rename withPkgsPath
Using __withPkgsPath is a way better name, because two underscores in
front of it commonly designate an internal attribute, which it is.

The callNetwork function now works in a similar vain as callMachine,
having a canonical way to specify the path to nixpkgs.

I've also changed the recursive search of machines to use the
__withPkgsPath attribute as a way to detect whether it's a machine.

Signed-off-by: aszlig <aszlig@nix.build>