pkgs/santander: Fix path to executable
The actual executable is libexec/travic-sign/travic-sign and not
libexec/travic-sign, so let's actually fix this in the native messaging
host config file.

This is mainly for the sake of OCD, because I'm not using Chromium
anymore and manually insert the path (I know, someday[TM] I will
automate this).

Signed-off-by: aszlig <aszlig@nix.build>