Remove my patched version of Gajim everywhere
The config.patch doesn't apply for Gajim 1.0 anymore anyway, so let's
throw everything away, including my custom config in order to start with
a new abomination.

With the new approach, I'm going to patch the configuration defaults
*directly* into Gajim, because one of the problems with the old approach
was that whenever specifics about a configuration value has changed, I
didn't get noticed by a patch failure.

So in the end the config I was ending up was a big mess.

I'm going to start this with a new unpatched version and someday get to
a patched version that I'm staisfied with... hopefully ;-)

Signed-off-by: aszlig <aszlig@nix.build>