tishtushi: Change to new partitioning scheme
I had to recover tishtushi while on the road because I have used zstd
compression, which GRUB doesn't support.

So instead of just adding a boot partition I decided to actually do
what I long wanted to do, which is to actually encrypt the disks, so the
new partitioning reflects that.

Signed-off-by: aszlig <aszlig@nix.build>