machines/shakti: Provide EDID information
The monitor reports back EDID via DVI but the HDMI->DVI adapter seems to
not handle that correctly. Also the monitor has weird resolutions, so
using the fallback modes provided by the kernel also doesn't seem to
work and the monitor stays blank.

While hardcoding the EDID information isn't a very good idea for this
machine in general (because it might be connected to a different
monitor), for now this is the easiest workaround because I don't have
access to that machine.

Signed-off-by: aszlig <aszlig@nix.build>