pkgs/vim: Add a few exceptions for sleuth
This has annoyed me for a while now, especially when writing Haskell
code.

Let's say if I have something like this:

foo :: Num a => a -> a
foo a =
    b
  where
    b = a + 1

Sleuth will detect that the indentation level is 2 spaces, but in
reality what I want is to have it at 4 spaces. So turning off Sleuth
will solve that problem, especially because we can still invoke it
manually.

Signed-off-by: aszlig <aszlig@nix.build>