pkgs/aszlig: Add gopass with custom patches
I only want gopass to use ASCII symbols. This is something I already
hated in pass, but I never changed it. By switching to gopass now, it's
a great opportunity to change it.

Second, I use "less -R" as a pager, which supports color so there really
is no need to disable colors before piping it to my pager.

Signed-off-by: aszlig <aszlig@nix.build>