pkgs/sandbox: Fix and pass through .desktop files
Since the introduction and move of a few packages to use the sandbox, we
no longer have XDG desktop entries, because the sandbox only creates
wrappers for all programs in $drv/bin.

This now also copies the XDG desktop files and replaces absolute paths
to binaries to refer to the sandboxed binaries.

I also modified the test to go through the XDG desktop file by default
so we can ensure that this works properly.

Signed-off-by: aszlig <aszlig@nix.build>