machines/titan: Replace fallbackDns
The original idea was to have globally reachable nameservers as
fallback, but really this isn't what I wanted. I want to have
nameservers additionally to the ones provided by DHCP, but in case the
one provided by DHCP does not deliver the resolve, the other primary
nameservers here will be queried instead.