games/stardew-valley: Add libGL to LD_LIBRARY_PATH
I haven't bisected this to know the exact commit that made this change
necessary, but it might be that SDL2 now no longer propagates libGL
anymore.

If this is the case for other games as well, I might want to do a
bisect, but for now let's just fix it.

Signed-off-by: aszlig <aszlig@nix.build>