pkgs: Fix passing through i686-linux package set
So far in the games namespace we have just used the callPackage_i686
function from <nixpkgs> instead of our augmented set.

If we just use packages that are available in <nixpkgs> everything is
fine, but as soon as we want to use one of our own packages for
i686-linux we can't simply do that.

One example is the override in the gog.albion derivation which just uses
buildSandbox from the main pkgs attribute set.

We now properly pass through the whole pkgsi686Linux set to the games
namespace, so we can drop that hack for Albion.

Signed-off-by: aszlig <aszlig@nix.build>