profiles/workstation: Refine zswap configuration
Since kernel 4.18 contains the zstd compression module in the crypto
API, it really makes sense to use that instead of the default (lzo) as
it seems to have lower CPU usage with higher compression ratios.

Another change I've made is to use z3fold for the pooling, so that 3
pages are cramped into one page of the pool. I did also have a look at
zsmalloc, but it seems to come with the cost of additional CPU usage.

Signed-off-by: aszlig <aszlig@nix.build>