games/bastion: Refactor Nix expression and update
With fetchHumbleBundle finally being able to fetch the game with my
account, I think it's time to update the Nix expression. :-)

The game's binary is basically nothing more than just a wrapped Mono, so
we can use directly Mono from <nixpkgs> instead of patching the wrapper.

I also stubbed out the symbols for libsteam_api.so, which might be a
good idea to provide for other games as well but in a more generic way.

Another thing that was needed solely because of my stubbornness was to
patch out the system() libc library call in libfmodex.so, because I
didn't want to add /bin/sh to the sandbox.

So in summary: The new expression now is sandboxed, doesn't use the
wrapper anymore and also has a desktop entry :-)

Unfortunately I haven't found a changelog online to see what has changed
between the 20120620 and 20161016.

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @layus