fetchHumbleBundle: Fix subproduct name collision
Every HumbleBundle purchase has a list of subproducts and those have
so-called 'machine_name' attributes which are unique. However in our
fetcher we used the 'human_name' as a unique key, which isn't really
unique and in case of bastion the soundtrack subproduct is has the
human_name "Bastion" and the actual game has the *same* name as well.

Using the machine_name as the unique key now solves the problem that
certain games couldn't be found if you for example also have the
soundtrack.

Signed-off-by: aszlig <aszlig@nix.build>