monogame-patcher: Add a replace-call subcommand
It's a very early version which I did for a Unity3d game and it
*desperately* needs a cleanup.

An example command line:

  monogame-patcher replace-call -i Assembly-CSharp.dll \
    'System.String UnityEngine.Application::get_dataPath()' \
    'System.String UnityEngine.Application::get_persistentDataPath()' \
    IniParser SaveMeta

Right now the replace-call subcommand has UnityEngine.Application
hardcoded, so this won't work for other types than
UnityEngine.Application.

However, I'm going to refactor the whole patcher very soon and add some
tests, so this whole mess will be cleaned up.

Signed-off-by: aszlig <aszlig@nix.build>