monogame-patcher: Switch to MSBuild
I really would have preferred a simple Makefile, but with that we can't
use buildDotnetPackage and we also need to take care of butchering the
dependencies manually.

So I moved everything to src/ and added a csproj file to clean up most
of the cruft and just use buildDotnetPackage with minimal attributes.

In addition to that I also added assembly info, so that the command line
help will show the proper author name, copyright, yaddayadda...

Signed-off-by: aszlig <aszlig@nix.build>