monogame-patcher: Restructure and add stub tests
Mainly this is so we can prepare for running unit tests, so we get the
latest version of NUnit and run the console test runner.

Currently there is only a dummy test which always succeeds, but it's
there so that we can fill out the boilerplate later.

I also moved the option definitions into a separate file so they don't
clutter up the main file.

Signed-off-by: aszlig <aszlig@nix.build>